+++ src/cli/bsky/front_matter.rs
@@ -1,0 +1,36 @@
+use color_eyre::Result;
+use serde::Deserialize;
+
+// +++
+// title = "Blue Sky Test Blog"
+// description = "A blog post to test the processing of blog posts for posting to Bluesky."
+// date = 2025-01-17
+// updated = 2025-01-16
+// draft = false
+//
+// [taxonomies]
+// topic = ["Technology"]
+// tags = ["bluesky", "testing", "test only", "ci"]
+// +++
+//
+
+#[derive(Default, Debug, Clone, Deserialize)]
+pub struct Taxonomies {
+    #[allow(dead_code)]
+    pub tags: Vec<String>,
+}
+
+#[derive(Default, Debug, Clone, Deserialize)]
+pub struct FrontMatter {
+    pub title: String,
+    pub description: String,
+    pub taxonomies: Taxonomies,
+    pub filename: Option<String>,
+}
+
+impl FrontMatter {
+    pub fn from_toml(toml: &str) -> Result<Self> {
+        let front_matter = toml::from_str::<FrontMatter>(toml)?;
+        Ok(front_matter)
+    }
+}